remove TODO about moving itable blocks instead of inodes
The ext2fs library already caches blocks when reading inodes.
Performace gains from moving inode table blocks seem unlikely,
compared to the current mechanism of reading inodes one by one,
which already profits from the library buffering.

If we move (read-write) blocks directly, we will still need to do
a second read to calculate the used inodes and folders stats for
each group. This doesn't seem faster than the existing approach.
If we want to avoid the second read, we could read the block,
parse the inodes in memory to update the stats, and write the block
to the new location. However, the parse step will complicate the code
too much (different inode sizes, endianness, etc...), and it is
uncertain it would be faster than the existing mechanism. Therefore,
we will keep the current approach that already profits from the
buffering done by the library, and has code easier to read and mantain.